+++ assets/overview-page.stories-ZfrCw1pb.js
@@ -1,0 +1,56 @@
+import{r as l,j as e}from"./iframe-DEKP6JnF.js";import{O as n}from"./OverviewPage-CtU5CLGa.js";import{d as u}from"./DiviExperienceHelperIcon-CEFrLmOf.js";import{A as d}from"./index-BhDAVdC-.js";import"./index-a6FQ02LQ.js";import"./index-BXvHdSVJ.js";import"./index-BLbezTN-.js";import"./PacAboutPage-y31k6Gyy.js";import"./index-DE_hdMpX.js";import"./ChangelogPage-B0VOtHiG.js";import{S as g}from"./index-TdP3Yctr.js";import"./Heading-CizVIwDG.js";import{T as i}from"./Text-CX_W7m7p.js";import"./FilledIcon-NMee4TIs.js";import"./index-7pMWvUSd.js";import"./index-asf2oUrF.js";import"./index-Bn31f6F_.js";import"./index-CwVHz0U5.js";import"./index-tAa_ZV3D.js";import"./index-DzXFNEVd.js";import"./index-DdNxV-yM.js";import"./index-sPWoe1m3.js";import"./index-onQU0KRF.js";import"./index-DVW88xY5.js";import"./index-CbbLOdae.js";import"./index-Dh5V6moa.js";import"./index-C4VI5kbb.js";import"./index-bB-AGrlo.js";import"./index-B_Zc13Eo.js";import"./index-sNzMoUEe.js";import"./index-_96lnEjV.js";import"./index-BQp2vdZi.js";import"./index-DUQan5Ow.js";import"./index-Cry99EDD.js";import"./index-DdoecMz6.js";import"./index-D_i5Qv6D.js";import"./index-DKzp5YS_.js";import"./index-BiwDJ9Cj.js";import"./index-B7piYs5c.js";import"./index-H4elOlt7.js";import"./index-BjdHKAtB.js";import"./index-GBesp_l2.js";import"./index-DcOqwMXe.js";import"./Label-BqFSMhGt.js";import{T as w}from"./TextLink-CPWnNgGs.js";import"./FeatureRequestPage-Csx3mbYE.js";import"./LicensePage-BCbySKrQ.js";import"./ChangelogIcon-jtPCg3Px.js";import"./CloseIcon-lpP-t_9C.js";import"./OpenIcon-xLe0hxFH.js";import"./DragIcon-2nbu_z8O.js";import"./HomeIcon-C-yey_9C.js";import"./RolesIcon-hKNPUeGt.js";import"./WidowFixerIcon-BwAdr2-g.js";import"./EmailIcon-BF8oZO78.js";import"./FaqIcon-BHZuKS3B.js";import"./KeyIcon-B0VgT2iW.js";import"./CacheIcon--3cyaBY_.js";import"./ImportExportIcon-CJbQJwfU.js";import"./InfoIcon-B79fggGH.js";import"./UnderlinedIcon--sc61Nb1.js";import"./LearningIcon-CqfRdQ89.js";import"./ResetIcon-CFYCsgpB.js";import"./UpdatesIcon-CgZ7Sh2a.js";import"./SearchIcon-DP9rDjnH.js";import"./SettingsIcon-DTvXUaxN.js";import"./SupportIcon-D6YLLDK5.js";import"./UsersIcon-8jGHjFr_.js";import"./WebsiteIcon-C6M4jHHV.js";import"./EyeIcon-BWAH84p3.js";import"./AddIcon-D67mipiB.js";import"./BackIcon-hod2ZtQg.js";import{S as v}from"./StopIcon-01QZEDtZ.js";import"./TrashIcon-Cp4sZOM2.js";import"./DropdownIcon-mVpGzupg.js";import"./ChevronRightIcon-BVKAzBaG.js";import"./PACLogo-KzghuyAC.js";import"./index-Dq7h7Pqt.js";import"./useOutsideToggle-Bmlz_kpv.js";import"./index-D7UbXefD.js";import"./index-CLd9FON8.js";import"./index-B3J4ROiz.js";import"./type-Bc9hys3F.js";const Ke={title:"PAC Universal Plugin UI/Pages/Overview Page",component:n,parameters:{pageLayout:"page"}},t=()=>{const[a,p]=l.useState(!0),m={avatar:"https://lh3.googleusercontent.com/ogw/AF2bZyg0NWhbUjvHEo5VTQIPmq2lgGyO4xDZeaWXPyHq5OknKQ=s32-c-mo",name:"Muhammad Touseef",greeting:"Good afternoon",pluginLogo:u,pluginDescription:"A huge collection of incredibly handy features and tools to make your life easier when working on your Divi website - startup actions, general utilities, interface hacks, custom settings, global styles, learning resources, maintenance tools, code editors, accessibility improvements, alt text fixes, media hacks, local fonts, system report sharing, import/export tools, and so much more!",pluginName:"Divi Assistant",pacPluginNewsApi:"https://dummyjson.com/c/f371-88b4-4d1e-b88e",onPacCardClick:()=>{console.log("PAC Card Clicked")},setCurrentTab:c=>{console.log("Current Tab Set To:",c)},docLink:"https://www.peeayecreative.com/docs/divi-assistant/",supportLink:"https://www.peeayecreative.com/support/",communityLink:"https://www.peeayecreative.com/community/space/divi-assistant/home",showBetaNotice:!0,betaNoticeLink:"https://www.peeayecreative.com/docs/divi-assistant/divi-5/",showReviewCard:a,onReviewCardDismiss:()=>{p(!1)},reviewCardLink:"https://www.elegantthemes.com/marketplace/divi-assistant/reviews/list/1",showLicenseAlert:!0,onLicenseAlertDismiss:()=>{console.log("License Alert Dismissed")},showCustomNotice:!0,customNoticeComponent:()=>e.jsx(d,{icon:v,type:"error",title:"Divi 5 Beta Compatbibility Notice",description:e.jsxs(g,{children:[e.jsx(i,{children:"Divi Contact Form Helper now includes full support for Divi 5 and can be installed on new Divi 5 beta sites. This update also includes full migration capability, allowing you to migrate your existing contact form settings from Divi 4 to Divi 5. Because Divi 5 is still in beta and the migration process is highly complex for a plugin of this nature, we strongly recommend testing all migrations on a staging site first. After migrating, please review your settings, especially custom field types such as file uploads and date pickers. We do not expect issues, but if anything unexpected occurs, please report it and we will improve the plugin as needed."}),e.jsxs(i,{children:["As a general note, Divi 5 continues to evolve, and future changes in the Divi 5 beta may require updates on our end. For the latest information, current status, and instructions for reporting issues, please see our ",e.jsx(w,{href:"https://www.peeayecreative.com/docs/divi-contact-form-helper/divi-5/",children:"Divi 5 Compatibility Guide."})]})]}),width:"stretch"})};return e.jsx(n,{...m})};t.__docgenInfo={description:"",methods:[],displayName:"Default"};var o,s,r;t.parameters={...t.parameters,docs:{...(o=t.parameters)==null?void 0:o.docs,source:{originalSource:`() => {
+  const [showReviewCard, setShowReviewCard] = useState(true);
+  const OverviewPageProps = {
+    avatar: "https://lh3.googleusercontent.com/ogw/AF2bZyg0NWhbUjvHEo5VTQIPmq2lgGyO4xDZeaWXPyHq5OknKQ=s32-c-mo",
+    name: "Muhammad Touseef",
+    greeting: "Good afternoon",
+    // pluginLogo: "https://dummyjson.com/image/74",
+    pluginLogo: DiviCustomFieldHeplerIcon,
+    pluginDescription: "A huge collection of incredibly handy features and tools to make your life easier when working on your Divi website - startup actions, general utilities, interface hacks, custom settings, global styles, learning resources, maintenance tools, code editors, accessibility improvements, alt text fixes, media hacks, local fonts, system report sharing, import/export tools, and so much more!",
+    pluginName: "Divi Assistant",
+    pacPluginNewsApi: "https://dummyjson.com/c/f371-88b4-4d1e-b88e",
+    onPacCardClick: () => {
+      console.log("PAC Card Clicked");
+    },
+    setCurrentTab: (tab: string) => {
+      console.log("Current Tab Set To:", tab);
+    },
+    docLink: "https://www.peeayecreative.com/docs/divi-assistant/",
+    supportLink: "https://www.peeayecreative.com/support/",
+    communityLink: "https://www.peeayecreative.com/community/space/divi-assistant/home",
+    showBetaNotice: true,
+    betaNoticeLink: "https://www.peeayecreative.com/docs/divi-assistant/divi-5/",
+    showReviewCard: showReviewCard,
+    onReviewCardDismiss: () => {
+      setShowReviewCard(false);
+    },
+    reviewCardLink: "https://www.elegantthemes.com/marketplace/divi-assistant/reviews/list/1",
+    showLicenseAlert: true,
+    onLicenseAlertDismiss: () => {
+      console.log("License Alert Dismissed");
+    },
+    showCustomNotice: true,
+    customNoticeComponent: () => <Alert icon={StopIcon} type="error" title={"Divi 5 Beta Compatbibility Notice"} description={<Stack>
+            <Text>
+              Divi Contact Form Helper now includes full support for Divi 5 and
+              can be installed on new Divi 5 beta sites. This update also
+              includes full migration capability, allowing you to migrate your
+              existing contact form settings from Divi 4 to Divi 5. Because Divi
+              5 is still in beta and the migration process is highly complex for
+              a plugin of this nature, we strongly recommend testing all
+              migrations on a staging site first. After migrating, please review
+              your settings, especially custom field types such as file uploads
+              and date pickers. We do not expect issues, but if anything
+              unexpected occurs, please report it and we will improve the plugin
+              as needed.
+            </Text>
+            <Text>
+              As a general note, Divi 5 continues to evolve, and future changes
+              in the Divi 5 beta may require updates on our end. For the latest
+              information, current status, and instructions for reporting
+              issues, please see our <TextLink href="https://www.peeayecreative.com/docs/divi-contact-form-helper/divi-5/">Divi 5 Compatibility Guide.</TextLink>
+            </Text>
+          </Stack>} width="stretch" />
+  };
+  return <OverviewPage {...OverviewPageProps} />;
+}`,...(r=(s=t.parameters)==null?void 0:s.docs)==null?void 0:r.source}}};const Me=["Default"];export{t as Default,Me as __namedExportsOrder,Ke as default};